Conjgrad for multiple threads not yet working

+++ CPP/comm_ctrl.cpp
@@ -3,7 +3,7 @@
 #include "operator_matrix.h"
 
 comm_ctrl::comm_ctrl(decomposition *d, operator_matrix *a,
-			double *rhs, double *rhs_up, double *U_up){			
+			double *rhs, double *rhs_up, double *U_up){
 	D=d; A=a; u_up=U_up; RHS=rhs; RHS_up=rhs_up;
 	init_neighbour_ranks();
 	init_group_behaviour();
@@ -26,15 +26,15 @@ void comm_ctrl::compile_solution(double *U){
 	MPI_Allreduce(MPI_IN_PLACE, U, D->get_N(), MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
 }
 
-void comm_ctrl::cumulate_dist_squared(double *dsq){	
+void comm_ctrl::cumulate_dist_squared(double *dsq){
 	MPI_Allreduce(MPI_IN_PLACE, dsq, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
 }
 
-void comm_ctrl::init_neighbour_ranks(){	
+void comm_ctrl::init_neighbour_ranks(){
   MPI_Comm_rank(MPI_COMM_WORLD, &myRank);
-	bottomRank= (D->get_myRank_y()>0) 									? myRank - D->get_N_procs_x() : -1; 
-	topRank= (D->get_myRank_y()<D->get_N_procs_y()-1) 	? myRank + D->get_N_procs_x() : -1; 
-	leftRank= (D->get_myRank_x()>0) 										? myRank-1 : -1; 
+	bottomRank= (D->get_myRank_y()>0) 									? myRank - D->get_N_procs_x() : -1;
+	topRank= (D->get_myRank_y()<D->get_N_procs_y()-1) 	? myRank + D->get_N_procs_x() : -1;
+	leftRank= (D->get_myRank_x()>0) 										? myRank-1 : -1;
 	rightRank= (D->get_myRank_x()<D->get_N_procs_x()-1) ? myRank+1 : -1;
 }
 
@@ -53,7 +53,7 @@ void comm_ctrl::send_updates(double *U){
 void comm_ctrl::send_update_toBottom(double *U){
 	if(bottomRank<0) return;
 	int *idx = D->get_index_global_bottom();
-	for(size_t i=0;i<D->get_myNx();++i){
+	for(int i=0;i<D->get_myNx();++i){
 		int j = idx[i];
 		u_up[j] = U[j];
 	}
@@ -63,7 +63,7 @@ void comm_ctrl::send_update_toBottom(double *U){
 void comm_ctrl::send_update_toTop(double *U){
 	if(topRank<0) return;
 	int *idx = D->get_index_global_top();
-	for(size_t i=0;i<D->get_myNx();++i){
+	for(int i=0;i<D->get_myNx();++i){
 		int j = idx[i];
 		u_up[j] = U[j];
 	}
@@ -73,7 +73,7 @@ void comm_ctrl::send_update_toTop(double *U){
 void comm_ctrl::send_update_toLeft(double *U){
 	if(leftRank<0) return;
 	int *idx = D->get_index_global_left();
-	for(size_t i=0;i<D->get_myNy();++i){
+	for(int i=0;i<D->get_myNy();++i){
 		int j = idx[i];
 		u_up[j] = U[j];
 	}
@@ -83,7 +83,7 @@ void comm_ctrl::send_update_toLeft(double *U){
 void comm_ctrl::send_update_toRight(double *U){
 	if(rightRank<0) return;
 	int *idx = D->get_index_global_right();
-	for(size_t i=0;i<D->get_myNy();++i){
+	for(int i=0;i<D->get_myNy();++i){
 		int j = idx[i];
 		u_up[j] = U[j];
 	}
@@ -102,9 +102,9 @@ void comm_ctrl::receive_update_fromBottom(){
 	MPI_Recv(u_up, D->get_N(), MPI_DOUBLE, bottomRank, 200, MPI_COMM_WORLD, &mpi_stat);
 	int *idx = D->get_index_global_bottom(),
 			offset = D->get_Nx();
-	for(size_t i=0;i<D->get_myNx();++i){
+	for(int i=0;i<D->get_myNx();++i){
 		int j = idx[i];
-		RHS_up[j] = RHS[j] - A->Cy() * u_up[j - offset]; 
+		RHS_up[j] = RHS[j] - A->Cy() * u_up[j - offset];
 	}
 }
 
@@ -113,9 +113,9 @@ void comm_ctrl::receive_update_fromTop(){
 	MPI_Recv(u_up, D->get_N(), MPI_DOUBLE, topRank, 100, MPI_COMM_WORLD, &mpi_stat);
 	int *idx = D->get_index_global_top(),
 			offset = D->get_Nx();
-	for(size_t i=0;i<D->get_myNx();++i){
+	for(int i=0;i<D->get_myNx();++i){
 		int j = idx[i];
-		RHS_up[j] = RHS[j] - A->Cy() * u_up[j + offset]; 
+		RHS_up[j] = RHS[j] - A->Cy() * u_up[j + offset];
 	}
 }
 
@@ -123,9 +123,9 @@ void comm_ctrl::receive_update_fromLeft(){
 	if(leftRank<0) return;
 	MPI_Recv(u_up, D->get_N(), MPI_DOUBLE, leftRank, 400, MPI_COMM_WORLD, &mpi_stat);
 	int *idx = D->get_index_global_left();
-	for(size_t i=0;i<D->get_myNy();++i){
+	for(int i=0;i<D->get_myNy();++i){
 		int j = idx[i];
-		RHS_up[j] = RHS[j] - A->Cx() * u_up[j-1]; 
+		RHS_up[j] = RHS[j] - A->Cx() * u_up[j-1];
 	}
 }
 
@@ -133,8 +133,8 @@ void comm_ctrl::receive_update_fromRight(){
 	if(rightRank<0) return;
 	MPI_Recv(u_up, D->get_N(), MPI_DOUBLE, rightRank, 300, MPI_COMM_WORLD, &mpi_stat);
 	int *idx = D->get_index_global_right();
-	for(size_t i=0;i<D->get_myNy();++i){
+	for(int i=0;i<D->get_myNy();++i){
 		int j = idx[i];
-		RHS_up[j] = RHS[j] - A->Cx() * u_up[j+1]; 
+		RHS_up[j] = RHS[j] - A->Cx() * u_up[j+1];
 	}
 }
